giveaway_bot/requirements.txt
pyTelegramBotAPI==4.14.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
aiohttp==3.9.1
